% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccopy.R
\name{cco}
\alias{cco}
\title{Copy commands only}
\usage{
cco()
}
\description{
Take the content of the system clipboard (assumed to contain text copied from
the R console), delete any lines of text not commencing with \code{"> "} or
\code{"+ "}, delete these characters from the beginning of all remaining
lines, and then write these remaining lines back to the clipboard.
}
\details{
This function should be run when the system clipboard contains text
  copied from the R console. \code{"> "} is the default console command
  prompt in R, and \code{"+ "} is the default console continuation prompt;
  deleting all lines of text that do not commence with these strings is
  intended to delete all lines that do not correspond to user commands (such
  as lines of output, or error/warning messages). From these remaining lines,
  the first two characters (either \code{"> "} and \code{"+ "}) are deleted.

  The resulting lines comprise executable R code, comprising only the user
  commands from the original console text. This is written to the clipboard
  in order that it may be pasted by the user.
}
\examples{
## Suppose the following (uncommented) lines are copied to the clipboard:
# > a <- 10
# > b <- 5
# > sum(a,
# + b)
# [1] 15
## The following line does this copying:
writeLines(c("> a <- 10", "> b <- 5", "> sum(a,", "+ b)", "[1] 15"),
  "clipboard")

## Copy commands only:
cco()

## Pasting from the clipboard gives the following (uncommented) lines:
# a <- 10
# b <- 5
# sum(a,
# b)
}
\seealso{
\code{\link{pco}}, \code{\link{rco}}
}

