% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccopy.R
\name{rco}
\alias{rco}
\title{Run commands only}
\usage{
rco(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{source}} (see below).}
}
\description{
Take the content of the system clipboard (assumed to contain text copied from
the R console), delete any lines of text not commencing with \code{"> "} or
\code{"+ "}, delete these characters from the beginning of all remaining
lines, and then execute these remaining lines in the console.
}
\details{
This function should be run when the system clipboard contains text
  copied from the R console. \code{"> "} is the default console command
  prompt in R, and \code{"+ "} is the default console continuation prompt;
  deleting all lines of text that do not commence with these strings is
  intended to delete all lines that do not correspond to user commands (such
  as lines of output, or error/warning messages). From these remaining lines,
  the first two characters (either \code{"> "} and \code{"+ "}) are deleted.

  The resulting lines comprise executable R code, comprising only the user
  commands from the original console text. This code is executed using the
  base function \code{\link{source}}, to which any arguments are passed (e.g.
  \code{echo = TRUE} causes each expression to be printed before evaluation).
}
\examples{
#' ## Suppose the following (uncommented) lines are copied to the clipboard:
# > a <- 10
# > b <- 5
# > sum(a,
# + b)
# [1] 15
## The following line does this copying:
writeLines(c("> a <- 10", "> b <- 5", "> sum(a,", "+ b)", "[1] 15"),
  "clipboard")

## Run commands only:
rco()

## Check assigned value of a (should be 10):
a
}

