% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccopy.R
\name{strippedConsoleText}
\alias{strippedConsoleText}
\title{Strip output, messages and command/continuation prompts from console text}
\usage{
strippedConsoleText(charVec)
}
\arguments{
\item{charVec}{A character vector (as might be read from the system
clipboard, where each element corresponds a new line of text). Intended to
contain text copied from the R console.}
}
\value{
A character vector obtained from \code{charVec} by deleting those
  elements not commencing with \code{"> "} or \code{"+ "}, and from the
  remaining strings deleting these first two characters.
}
\description{
Utility function. For a vector of character strings, each string that does
not commence either with \code{"> "} (the console command prompt) or \code{"+
"} (the console continuation prompt) is removed. These prompt characters are
then deleted from each of the remaining strings.
}
\keyword{internal}

